Move test to appio
